Increase Item uid from 255 to 4096 bytes
Slashdot regularly has stories with uids over 255 bytes, which breaks
the 'item already in database' logic (it compares the truncated version
from the database to the untruncated version from the web).  As there is
no apparent performance or memory implication for increasing the size of
the varchar, bump it significantly.