Keep selected entry title dark
With auto_mark_read turned on, the entry, and the title, will have the
unread class removed on selecting/opening.  This makes it very hard to
read the title of the article.  Instead, keep the same unread dark color
as long as the entry is selected.
(cherry picked from commit 0469e92a341e7b99cbd28b7bb14e9e9cff324749)