Implement swipe navigation in fullscreen mode on mobile
Integrates the touchSwipe jquery plugin to provide swipe events and uses
some animation to provide feedback to the user during swipes.

On a stock Galaxy Nexus (4.2.1) the animation can be a bit choppy, but I
think it's due to jquery replacing the contents of the fullscreen div in
the middle of the animation rather than the animation itself.